refactor(auth): extract remote session helpers
